Frontend: React app — marshal form, status page, and full coordinator/committee dashboard
Implements spec sections 6-7, 10, 16:
- Brand theme (navy/orange/parchment, system fonts, buttons, badges, cards)
- Logo (pin + flag) and favicon
- Public marshal application form: 6 sections, validation, live cost totals,
  30s localStorage draft autosave, confirmation page, licence attach
- Marshal status page with licence upload and BACS payment details
- Admin: JWT auth context, event selector, sidebar layout with role gating
- Dashboard, applications table (filters/sort/bulk/detail slide-over),
  scheduling grid + ORA auto-assign preview/commit + lock + daily roster,
  communications (send + log tabs), payments, reports, committee view
- Axios client with auth interceptor and authenticated CSV downloads

Also fixes an integer-cast bug in the licence verification SQL.

+++ server/routes/applications.js
@@ -150,7 +150,7 @@ router.put('/admin/applications/:id/verify-licence', requireAuth, requireCoordin
     const { rows } = await db.query(
       `UPDATE marshals
        SET licence_verified = $1,
-           licence_verified_by = CASE WHEN $1 THEN $2 ELSE NULL END,
+           licence_verified_by = CASE WHEN $1 THEN $2::int ELSE NULL END,
            licence_verified_at = CASE WHEN $1 THEN NOW() ELSE NULL END,
            updated_at = NOW()
        WHERE id = $3 RETURNING licence_verified, licence_verified_at`,
